More optimizations for tie-dye mode

+++ shaders/julia.frag
@@ -15,20 +15,20 @@ uniform float flux;
 
 void main() {
     vec2 uv = 0.3 * vPos.xy;
-    float t = (u_time * 0.01) + onset;
+    mediump float t = (u_time * 0.01) + onset;
 
-    float k = cos(t);
-    float l = sin(t);
-    float s = 0.2 + (onset/10.0);
+    mediump float k = cos(t);
+    mediump float l = sin(t);
+    mediump float s = 0.2 + (onset/10.0);
 
     // XXX simplify back to shadertoy example
-    for(int i=0; i<32; ++i) {
+    for(int i = 0; i < 32; ++i) {
         uv  = abs(uv) - s;//-onset;    // Mirror
         uv *= mat2(k,-l,l,k); // Rotate
         s  *= .95156;///(t+1);         // Scale
     }
 
-    float x = .5 + .5 * cos(6.28318 * (40.0 * length(uv)));
+    mediump float x = .5 + .5 * cos(6.28318 * (40.0 * length(uv)));
     fragColor = .5 + .5 * cos(6.28318 * (40.0 * length(uv)) * vec4(-1,2 + (u_time / 500.0), 3 + flux, 1));
 }
